fix(parent): #EVAL-524 remove structure filter on /enfant endpoint (due to mobil use bug).
