+++ src/main/java/fr/openent/competences/security/IsRelative.java
@@ -1,0 +1,15 @@
+package fr.openent.competences.security;
+
+import fr.openent.competences.constants.Field;
+import fr.wseduc.webutils.http.Binding;
+import io.vertx.core.Handler;
+import io.vertx.core.http.HttpServerRequest;
+import org.entcore.common.http.filter.ResourcesProvider;
+import org.entcore.common.user.UserInfos;
+
+public class IsRelative implements ResourcesProvider {
+    @Override
+    public void authorize(HttpServerRequest request, Binding binding, UserInfos user, Handler<Boolean> handler) {
+        handler.handle(user.getType().equals(Field.RELATIVE));
+    }
+}
