--- src/test/java/fr/openent/competences/security/AccessStructureIsParentTest.java
@@ -1,99 +1,0 @@
-package fr.openent.competences.security;
-
-import fr.wseduc.webutils.http.Binding;
-import io.netty.handler.codec.http.DefaultHttpHeaders;
-import io.vertx.core.MultiMap;
-import io.vertx.core.http.HttpServerRequest;
-import io.vertx.core.http.impl.HeadersAdaptor;
-import io.vertx.ext.unit.Async;
-import io.vertx.ext.unit.TestContext;
-import io.vertx.ext.unit.junit.VertxUnitRunner;
-import org.entcore.common.user.UserInfos;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mockito;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@RunWith(VertxUnitRunner.class)
-public class AccessStructureIsParentTest {
-    AccessStructureIsParent access;
-    HttpServerRequest request;
-    Binding binding;
-    MultiMap map;
-    List<String> structures;
-    UserInfos user;
-
-    @Before
-    public void setUp() throws NoSuchFieldException {
-        access = new AccessStructureIsParent();
-        request = Mockito.mock(HttpServerRequest.class);
-        binding = Mockito.mock(Binding.class);
-        map = Mockito.spy(new HeadersAdaptor(new DefaultHttpHeaders()));
-        structures = new ArrayList<>();
-        user = new UserInfos();
-    }
-
-    @Test
-    public void testAuthorize(TestContext ctx){
-        map.set("id_structure", "9af51dc6-ead0-4edb-8978-da14a3e9f49a");
-        Mockito.doReturn(map).when(request).params();
-        structures.add("9af51dc6-ead0-4edb-8978-da14a3e9f49a");
-        user.setStructures(structures);
-        user.setType("Relative");
-        Async async = ctx.async();
-        access.authorize(request, binding, user, result -> {
-            ctx.assertEquals(true, result);
-            async.complete();
-        });
-        async.awaitSuccess(10000);
-    }
-
-    @Test
-    public void testBadUserType(TestContext ctx){
-        map.set("id_structure", "9af51dc6-ead0-4edb-8978-da14a3e9f49a");
-        Mockito.doReturn(map).when(request).params();
-        structures.add("9af51dc6-ead0-4edb-8978-da14a3e9f49a");
-        user.setStructures(structures);
-        user.setType("Teacher");
-        Async async = ctx.async();
-        access.authorize(request, binding, user, result -> {
-            ctx.assertEquals(false, result);
-            async.complete();
-        });
-        async.awaitSuccess(10000);
-    }
-
-    @Test
-    public void testBadStructure(TestContext ctx){
-        map.set("id_structure", "9af51dc6-ead0-4edb-8978-da14a3e9f49a");
-        Mockito.doReturn(map).when(request).params();
-        structures.add("aaaaaa");
-        user.setStructures(structures);
-        user.setType("Relative");
-        Async async = ctx.async();
-        access.authorize(request, binding, user, result -> {
-            ctx.assertEquals(false, result);
-            async.complete();
-        });
-        async.awaitSuccess(10000);
-    }
-
-
-    @Test
-    public void testBadStructureBadType(TestContext ctx){
-        map.set("id_structure", "9af51dc6-ead0-4edb-8978-da14a3e9f49a");
-        Mockito.doReturn(map).when(request).params();
-        structures.add("aaaaaa");
-        user.setStructures(structures);
-        user.setType("Teacher");
-        Async async = ctx.async();
-        access.authorize(request, binding, user, result -> {
-            ctx.assertEquals(false, result);
-            async.complete();
-        });
-        async.awaitSuccess(10000);
-    }
-}
